Extract common connector code to shared file
