+++ src/connectors/webauthn-fallback.ts
@@ -1,4 +1,5 @@
 import { getQsParam } from './common';
+import { b64Decode, buildDataString } from './common-webauthn';
 
 // tslint:disable-next-line
 require('./webauthn-fallback.scss');
@@ -72,7 +73,7 @@ function start() {
     initWebAuthn(json);
 }
 
-function initWebAuthn(obj: any) {
+async function initWebAuthn(obj: any) {
     const challenge = obj.challenge.replace(/-/g, '+').replace(/_/g, '/');
     obj.challenge = Uint8Array.from(atob(challenge), c => c.charCodeAt(0));
 
@@ -82,39 +83,22 @@ function initWebAuthn(obj: any) {
         listItem.id = Uint8Array.from(atob(fixedId), c => c.charCodeAt(0));
     });
 
-    navigator.credentials.get({ publicKey: obj })
-        .then((assertedCredential: PublicKeyCredential) => {
-            if (sentSuccess) {
-                return;
-            }
+    try {
+        const assertedCredential = await navigator.credentials.get({ publicKey: obj }) as PublicKeyCredential;
+        
+        if (sentSuccess) {
+            return;
+        }
 
-            const response = assertedCredential.response as AuthenticatorAssertionResponse;
+        const dataString = buildDataString(assertedCredential);
+        const remember = (document.getElementById('remember') as HTMLInputElement).checked;
+        window.postMessage({ command: 'webAuthnResult', data: dataString, remember: remember }, '*');
 
-            const authData = new Uint8Array(response.authenticatorData);
-            const clientDataJSON = new Uint8Array(response.clientDataJSON);
-            const rawId = new Uint8Array(assertedCredential.rawId);
-            const sig = new Uint8Array(response.signature);
-
-            const data = {
-                id: assertedCredential.id,
-                rawId: coerceToBase64Url(rawId),
-                type: assertedCredential.type,
-                extensions: assertedCredential.getClientExtensionResults(),
-                response: {
-                    authenticatorData: coerceToBase64Url(authData),
-                    clientDataJson: coerceToBase64Url(clientDataJSON),
-                    signature: coerceToBase64Url(sig),
-                },
-            };
-
-            const dataString = JSON.stringify(data);
-            const remember = (document.getElementById('remember') as HTMLInputElement).checked;
-            window.postMessage({ command: 'webAuthnResult', data: dataString, remember: remember }, '*');
-
-            sentSuccess = true;
-            success(translate('webAuthnSuccess'));
-        })
-        .catch(err => error(err));
+        sentSuccess = true;
+        success(translate('webAuthnSuccess'));
+    } catch (err) {
+        error(err);
+    }
 }
 
 function error(message: string) {
@@ -131,44 +115,4 @@ function success(message: string) {
     el.innerHTML = message;
     el.classList.add('alert');
     el.classList.add('alert-success');
-}
-
-function b64Decode(str: string) {
-    return decodeURIComponent(Array.prototype.map.call(atob(str), (c: string) => {
-        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-    }).join(''));
-}
-
-// From https://github.com/abergs/fido2-net-lib/blob/b487a1d47373ea18cd752b4988f7262035b7b54e/Demo/wwwroot/js/helpers.js#L34
-// License: https://github.com/abergs/fido2-net-lib/blob/master/LICENSE.txt
-function coerceToBase64Url(thing: any) {
-    // Array or ArrayBuffer to Uint8Array
-    if (Array.isArray(thing)) {
-        thing = Uint8Array.from(thing);
-    }
-
-    if (thing instanceof ArrayBuffer) {
-        thing = new Uint8Array(thing);
-    }
-
-    // Uint8Array to base64
-    if (thing instanceof Uint8Array) {
-        let str = '';
-        const len = thing.byteLength;
-
-        for (let i = 0; i < len; i++) {
-            str += String.fromCharCode(thing[i]);
-        }
-        thing = window.btoa(str);
-    }
-
-    if (typeof thing !== 'string') {
-        throw new Error('could not coerce to string');
-    }
-
-    // base64 to base64url
-    // NOTE: "=" at the end of challenge is optional, strip it off here
-    thing = thing.replace(/\+/g, '-').replace(/\//g, '_').replace(/=*$/g, '');
-
-    return thing;
 }
