+++ src/connectors/webauthn.ts
@@ -1,4 +1,5 @@
 import { getQsParam } from './common';
+import { b64Decode, buildDataString } from './common-webauthn';
 
 document.addEventListener('DOMContentLoaded', () => {
     init();
@@ -6,7 +7,6 @@ document.addEventListener('DOMContentLoaded', () => {
 
 let parentUrl: string = null;
 let parentOrigin: string = null;
-let version: number = null;
 let stopWebAuthn = false;
 let sentSuccess = false;
 
@@ -39,15 +39,8 @@ function start() {
         parentOrigin = new URL(parentUrl).origin;
     }
 
-    const versionQs = getQsParam('v');
-    if (!versionQs) {
-        error('No version.');
-        return;
-    }
-
     let json: any;
     try {
-        version = parseInt(versionQs, 10);
         const jsonString = b64Decode(data);
         json = JSON.parse(jsonString);
     }
@@ -103,26 +96,7 @@ function success(assertedCredential: PublicKeyCredential) {
         return;
     }
 
-    const response = assertedCredential.response as AuthenticatorAssertionResponse;
-
-    const authData = new Uint8Array(response.authenticatorData);
-    const clientDataJSON = new Uint8Array(response.clientDataJSON);
-    const rawId = new Uint8Array(assertedCredential.rawId);
-    const sig = new Uint8Array(response.signature);
-
-    const data = {
-        id: assertedCredential.id,
-        rawId: coerceToBase64Url(rawId),
-        type: assertedCredential.type,
-        extensions: assertedCredential.getClientExtensionResults(),
-        response: {
-            authenticatorData: coerceToBase64Url(authData),
-            clientDataJson: coerceToBase64Url(clientDataJSON),
-            signature: coerceToBase64Url(sig),
-        },
-    };
-
-    const dataString = JSON.stringify(data);
+    const dataString = buildDataString(assertedCredential);
     parent.postMessage('success|' + dataString, parentUrl);
     sentSuccess = true;
 }
@@ -131,42 +105,3 @@ function info(message: string) {
     parent.postMessage('info|' + message, parentUrl);
 }
 
-function b64Decode(str: string) {
-    return decodeURIComponent(Array.prototype.map.call(atob(str), (c: string) => {
-        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-    }).join(''));
-}
-
-// From https://github.com/abergs/fido2-net-lib/blob/b487a1d47373ea18cd752b4988f7262035b7b54e/Demo/wwwroot/js/helpers.js#L34
-// License: https://github.com/abergs/fido2-net-lib/blob/master/LICENSE.txt
-function coerceToBase64Url(thing: any) {
-    // Array or ArrayBuffer to Uint8Array
-    if (Array.isArray(thing)) {
-        thing = Uint8Array.from(thing);
-    }
-
-    if (thing instanceof ArrayBuffer) {
-        thing = new Uint8Array(thing);
-    }
-
-    // Uint8Array to base64
-    if (thing instanceof Uint8Array) {
-        let str = '';
-        const len = thing.byteLength;
-
-        for (let i = 0; i < len; i++) {
-            str += String.fromCharCode(thing[i]);
-        }
-        thing = window.btoa(str);
-    }
-
-    if (typeof thing !== 'string') {
-        throw new Error('could not coerce to string');
-    }
-
-    // base64 to base64url
-    // NOTE: "=" at the end of challenge is optional, strip it off here
-    thing = thing.replace(/\+/g, '-').replace(/\//g, '_').replace(/=*$/g, '');
-
-    return thing;
-}
